Creates UIHelper
