+++ GitHubFollowers/Screens/FollowersListViewController.swift
@@ -9,9 +9,7 @@ import UIKit
 
 class FollowersListViewController: UIViewController {
 
-    enum Section {
-        case main
-    }
+    enum Section { case main }
 
     var username: String?
     var followers: [Follower] = []
@@ -39,24 +37,10 @@ class FollowersListViewController: UIViewController {
     }
 
     func configureCollectionview() {
-        collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: createThreeColumnFlowLayout())
+        collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: UIHelper.createThreeColumnFlowLayout(in: view))
         view.addSubview(collectionView)
         collectionView.backgroundColor = .systemBackground
         collectionView.register(FollowerCell.self, forCellWithReuseIdentifier: FollowerCell.reuseID)
-    }
-
-    func createThreeColumnFlowLayout() -> UICollectionViewFlowLayout {
-        let width = view.bounds.width
-        let padding: CGFloat = 12
-        let minimumItemSpasing: CGFloat = 10
-        let availableWidth = width - (padding * 2) - (minimumItemSpasing * 2)
-        let itemWidth = availableWidth / 3
-
-        let flowLayout = UICollectionViewFlowLayout()
-        flowLayout.sectionInset = UIEdgeInsets(top: padding, left: padding, bottom: padding, right: padding)
-        flowLayout.itemSize = CGSize(width: itemWidth, height: itemWidth + 40)
-
-        return flowLayout
     }
 
     func getfollowers() {
