+++ GitHubFollowers/Utilities/UIHelper.swift
@@ -1,0 +1,25 @@
+//
+//  UIHelper.swift
+//  GitHubFollowers
+//
+//  Created by Valeriia Zakharova on 02.03.2021.
+//
+
+import UIKit
+
+struct UIHelper {
+
+    static func createThreeColumnFlowLayout(in view: UIView) -> UICollectionViewFlowLayout {
+        let width = view.bounds.width
+        let padding: CGFloat = 12
+        let minimumItemSpasing: CGFloat = 10
+        let availableWidth = width - (padding * 2) - (minimumItemSpasing * 2)
+        let itemWidth = availableWidth / 3
+
+        let flowLayout = UICollectionViewFlowLayout()
+        flowLayout.sectionInset = UIEdgeInsets(top: padding, left: padding, bottom: padding, right: padding)
+        flowLayout.itemSize = CGSize(width: itemWidth, height: itemWidth + 40)
+
+        return flowLayout
+    }
+}
